A Simple C Program
This program can read the full name, an 11-digit integer, float, and 2-digit decimal numbers in appropriate variables. Also, for the 11-digit number, it can divide the first 6 figure of the number by the last 5 numbers. It is able to convert numbers read from keyboard from decimal to the octal and hexadecimal format. In addition, it can convert degrees Calsius to degrees Fahrenheit and degrees Kelvin. When entering real numbers a, b, c,it can judge whether there is any roots and solve the problem by the way.